Merge dev to main

+++ src/main/java/com/brewstack/api/dto/RecipeIngredientDTO.java
@@ -1,7 +1,9 @@
 package com.brewstack.api.dto;
 
+import java.math.BigDecimal;
+
 public record RecipeIngredientDTO(
         String ingredientName,
         String unit,
-        Double quantityRequired
+        BigDecimal quantityRequired
 ) {}
